refactor: BaseEntity , DelYN editor
