+++ .run/NoExitApplication.run.xml
@@ -5,8 +5,8 @@
       <env name="MONGODB_USERNAME" value="root" />
       <env name="MONGODB_PASSWORD" value="1234" />
       <env name="MONGODB_PORT" value="8080" />
-      <env name="MAIL_USERNAME" value="no-reply@gmail.com" />
-      <env name="MAIL_PASSWORD" value="jvdd poxa ikuk ubya " />
+      <env name="MAIL_USERNAME" value="psh7492@gmail.com" />
+      <env name="MAIL_PASSWORD" value="qnda xcpo bnnw diun " />
       <env name="MAIL_HOST" value="smtp.gmail.com" />
       <env name="MAIL_PORT" value="587" />
       <env name="REDIS_HOST" value="localhost" />
