+++ src/main/java/com/E1i3/NoExit/domain/board/dto/BoardCreateReqDto.java
@@ -2,7 +2,6 @@ package com.E1i3.NoExit.domain.board.dto;
 
 import com.E1i3.NoExit.domain.board.domain.Board;
 import com.E1i3.NoExit.domain.board.domain.BoardType;
-import com.E1i3.NoExit.domain.board.domain.DelYN;
 import com.E1i3.NoExit.domain.member.domain.Member;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
