+++ src/main/java/com/E1i3/NoExit/domain/comment/domain/Comment.java
@@ -6,6 +6,8 @@ import com.E1i3.NoExit.domain.board.dto.BoardListResDto;
 import com.E1i3.NoExit.domain.board.dto.BoardUpdateReqDto;
 import com.E1i3.NoExit.domain.comment.dto.CommentListResDto;
 import com.E1i3.NoExit.domain.comment.dto.CommentUpdateReqDto;
+import com.E1i3.NoExit.domain.common.domain.BaseTimeEntity;
+import com.E1i3.NoExit.domain.common.domain.DelYN;
 import com.E1i3.NoExit.domain.member.domain.Member;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,7 +24,7 @@ import java.time.LocalDateTime;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Comment {
+public class Comment extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -41,12 +43,6 @@ public class Comment {
 
     private int dislikes; // 싫어요 수
 
-    @CreationTimestamp
-    private LocalDateTime createdTime; // 작성시간
-
-    @UpdateTimestamp
-    private LocalDateTime updatedTime; // 수정시간
-
     @Enumerated(EnumType.STRING)
     @Builder.Default
     private DelYN delYN = DelYN.N; // 삭제 여부
@@ -59,8 +55,8 @@ public class Comment {
                 .content(this.content)
                 .likes(this.likes)
                 .dislikes(this.dislikes)
-                .createdTime(this.createdTime)
-                .updatedTime(this.updatedTime)
+                .createdTime(this.getCreatedTime())
+                .updatedTime(this.getUpdateTime())
                 .build();
 
         return commentListResDto;
