+++ src/main/java/com/E1i3/NoExit/domain/comment/dto/CommentCreateReqDto.java
@@ -2,7 +2,6 @@ package com.E1i3.NoExit.domain.comment.dto;
 
 import com.E1i3.NoExit.domain.board.domain.Board;
 import com.E1i3.NoExit.domain.comment.domain.Comment;
-import com.E1i3.NoExit.domain.comment.domain.DelYN;
 import com.E1i3.NoExit.domain.member.domain.Member;
 import lombok.AllArgsConstructor;
 import lombok.Data;
