+++ src/main/java/com/E1i3/NoExit/domain/common/domain/BaseTimeEntity.java
@@ -4,12 +4,14 @@ import java.time.LocalDateTime;
 
 import javax.persistence.MappedSuperclass;
 
+import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import lombok.Getter;
 
-@Getter
+
+@Data
 @MappedSuperclass
 public abstract class BaseTimeEntity {
 	@CreationTimestamp
