+++ src/main/java/com/E1i3/NoExit/domain/common/domain/DelYN.java
@@ -1,0 +1,5 @@
+package com.E1i3.NoExit.domain.common.domain;
+
+public enum DelYN {
+    Y,N
+}
