--- src/main/java/com/E1i3/NoExit/domain/findboard/domain/DelYn.java
@@ -1,5 +1,0 @@
-package com.E1i3.NoExit.domain.findboard.domain;
-
-public enum DelYn {
-    Y,N
-}
