+++ src/main/java/com/E1i3/NoExit/domain/findboard/game/domain/AgeLimit.java
@@ -1,0 +1,5 @@
+package com.E1i3.NoExit.domain.findboard.game.domain;
+
+public enum AgeLimit {
+    ADULT, STUDENT
+}
