+++ src/main/java/com/E1i3/NoExit/domain/findboard/game/domain/Difficult.java
@@ -1,4 +1,4 @@
-package com.E1i3.NoExit.domain.game.domain;
+package com.E1i3.NoExit.domain.findboard.game.domain;
 
 public enum Difficult {
     ONE, TWO, THREE, FOUR, FIVE;
