+++ src/main/java/com/E1i3/NoExit/domain/findboard/game/repository/GameRepository.java
@@ -1,11 +1,8 @@
-package com.E1i3.NoExit.domain.game.repository;
+package com.E1i3.NoExit.domain.findboard.game.repository;
 
-import com.E1i3.NoExit.domain.game.domain.Game;
-import com.E1i3.NoExit.domain.reservation.domain.Reservation;
+import com.E1i3.NoExit.domain.findboard.game.domain.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface GameRepository extends JpaRepository<Game, Long> {
