--- src/main/java/com/E1i3/NoExit/domain/game/domain/AgeLimit.java
@@ -1,5 +1,0 @@
-package com.E1i3.NoExit.domain.game.domain;
-
-public enum AgeLimit {
-    ADULT, STUDENT
-}
