--- src/main/java/com/E1i3/NoExit/domain/member/domain/DelYN.java
@@ -1,5 +1,0 @@
-package com.E1i3.NoExit.domain.member.domain;
-
-public enum DelYN {
-	Y,N
-}
