+++ src/main/java/com/E1i3/NoExit/domain/member/service/MemberService.java
@@ -1,7 +1,5 @@
 package com.E1i3.NoExit.domain.member.service;
 
-import java.time.Duration;
-
 import javax.persistence.EntityExistsException;
 import javax.persistence.EntityNotFoundException;
 
@@ -20,7 +18,7 @@ import com.E1i3.NoExit.domain.member.dto.MemberListResDto;
 import com.E1i3.NoExit.domain.member.dto.MemberSaveReqDto;
 import com.E1i3.NoExit.domain.member.dto.MemberUpdateDto;
 import com.E1i3.NoExit.domain.member.repository.MemberRepository;
-import com.E1i3.NoExit.domain.owner.domain.Owner;
+import com.E1i3.NoExit.domain.owner.controller.domain.Owner;
 import com.E1i3.NoExit.domain.owner.repository.OwnerRepository;
 
 @Service
