+++ src/main/java/com/E1i3/NoExit/domain/owner/controller/domain/Owner.java
@@ -1,4 +1,4 @@
-package com.E1i3.NoExit.domain.owner.domain;
+package com.E1i3.NoExit.domain.owner.controller.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -9,10 +9,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 import com.E1i3.NoExit.domain.common.domain.BaseTimeEntity;
-import com.E1i3.NoExit.domain.member.domain.DelYN;
+import com.E1i3.NoExit.domain.common.domain.DelYN;
 import com.E1i3.NoExit.domain.member.domain.Role;
 import com.E1i3.NoExit.domain.owner.dto.OwnerListResDto;
-import com.E1i3.NoExit.domain.owner.dto.OwnerSaveReqDto;
 import com.E1i3.NoExit.domain.owner.dto.OwnerUpdateDto;
 
 import lombok.AllArgsConstructor;
