+++ src/main/java/com/E1i3/NoExit/domain/owner/repository/OwnerRepository.java
@@ -5,8 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.E1i3.NoExit.domain.member.domain.Member;
-import com.E1i3.NoExit.domain.owner.domain.Owner;
+import com.E1i3.NoExit.domain.owner.controller.domain.Owner;
 
 @Repository
 public interface OwnerRepository extends JpaRepository<Owner, Long> {
