+++ src/main/java/com/E1i3/NoExit/domain/review/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
 package com.E1i3.NoExit.domain.review.repository;
 
-import com.E1i3.NoExit.domain.member.domain.DelYN;
+import com.E1i3.NoExit.domain.common.domain.DelYN;
 import com.E1i3.NoExit.domain.member.domain.Member;
 import com.E1i3.NoExit.domain.member.repository.MemberRepository;
 import com.E1i3.NoExit.domain.reservation.domain.Reservation;
@@ -16,6 +16,9 @@ import java.util.Optional;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    List<Review> findByMemberAndDelYn(Member member, DelYN delYn);
-    Optional<Review> findByReservationAndDelYn(Reservation reservation, DelYN delYn);
+
+    List<Review> findByMemberAndDelYN(Member member, DelYN delYn);
+    Optional<Review> findByReservationAndDelYN(Reservation reservation, DelYN delYn);
+
+
 }
